Added test for rx_data processing
